Fixing presubmit
Exact reason this started failing spontaneously is unclear. However,
the failing behavior is transitional anyway and will be removed once the
Travel app is fully integrated with Syncbase watch, and the last set of
changes I made to the mock made the 'memory' flag on the update even
unwarranted anyway.

Change-Id: Ife3e84ccc3553a88a1523a64c5a067f7a052d8f2